Update all dependencies as much as possible
`sass-lint` is unmaintained and has not been updated in four years. As
such, it's about time to sunset this engine but I want to get it in
a clean state before I do.

This updates all dependencies as much as I can update them. There are
still a few security advisories, but I can't do much about them without
switching to Yarn for its custom resolutions.